# hosting-resources-service/cmd/server/rest/gen/api.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreatePoolRequest defines model for CreatePoolRequest.
type CreatePoolRequest struct {
	CpuCores int    `json:"cpuCores"`
	DiskGb   int    `json:"diskGb"`
	IpCount  int    `json:"ipCount"`
	Name     string `json:"name"`
	RamMb    int    `json:"ramMb"`
}

// Link defines model for Link.
type Link struct {
	Href string `json:"href"`

	// Templated Указывает, является ли 'href' URI-шаблоном (RFC 6570).
	Templated *bool `json:"templated,omitempty"`
}

// Links Контейнер для гипермедиа-ссылок.
type Links map[string]Link

// PageMetadata Информация о пагинации
type PageMetadata struct {
	// HasNextPage Есть ли следующая страница
	HasNextPage bool `json:"hasNextPage"`

	// HasPreviousPage Есть ли предыдущая страница
	HasPreviousPage bool `json:"hasPreviousPage"`

	// Number Текущий номер страницы
	Number int `json:"number"`

	// Size Размер страницы
	Size int `json:"size"`

	// TotalElements Общее количество элементов
	TotalElements int `json:"totalElements"`

	// TotalPages Общее количество страниц
	TotalPages int `json:"totalPages"`
}

// Pool defines model for Pool.
type Pool struct {
	// UnderscoreLinks Контейнер для гипермедиа-ссылок.
	UnderscoreLinks Links              `json:"_links"`
	Id              openapi_types.UUID `json:"id"`
	Name            string             `json:"name"`
	Resources       Resource           `json:"resources"`
}

// PoolCollectionResponse defines model for PoolCollectionResponse.
type PoolCollectionResponse struct {
	UnderscoreEmbedded struct {
		Pools []Pool `json:"pools"`
	} `json:"_embedded"`

	// UnderscoreLinks Контейнер для гипермедиа-ссылок.
	UnderscoreLinks Links `json:"_links"`

	// Page Информация о пагинации
	Page PageMetadata `json:"page"`
}

// Resource defines model for Resource.
type Resource struct {
	CpuCores int `json:"cpuCores"`
	DiskGb   int `json:"diskGb"`
	IpCount  int `json:"ipCount"`
	RamMb    int `json:"ramMb"`
}

// RootResource defines model for RootResource.
type RootResource struct {
	// UnderscoreLinks Контейнер для гипермедиа-ссылок.
	UnderscoreLinks Links `json:"_links"`
}

// StatusResponse defines model for StatusResponse.
type StatusResponse struct {
	Message string `json:"message"`
}

// Page defines model for Page.
type Page = int

// PageSize defines model for PageSize.
type PageSize = int

// BadRequest defines model for BadRequest.
type BadRequest = StatusResponse

// NotFound defines model for NotFound.
type NotFound = StatusResponse

// ListPoolsParams defines parameters for ListPools.
type ListPoolsParams struct {
	// Page Номер запрашиваемой страницы
	Page *Page `form:"page,omitempty" json:"page,omitempty"`

	// PageSize Количество элементов на странице.
	PageSize *PageSize `form:"pageSize,omitempty" json:"pageSize,omitempty"`
}

// CreatePoolJSONRequestBody defines body for CreatePool for application/json ContentType.
type CreatePoolJSONRequestBody = CreatePoolRequest

// AddResourcesJSONRequestBody defines body for AddResources for application/json ContentType.
type AddResourcesJSONRequestBody = Resource

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Точка входа (Root)
	// (GET /)
	GetRoot(w http.ResponseWriter, r *http.Request)
	// Получить список пулов
	// (GET /pools)
	ListPools(w http.ResponseWriter, r *http.Request, params ListPoolsParams)
	// Создание нового пула
	// (POST /pools)
	CreatePool(w http.ResponseWriter, r *http.Request)
	// Добавить ресурсы в пул
	// (POST /pools/{poolId}/resources)
	AddResources(w http.ResponseWriter, r *http.Request, poolId openapi_types.UUID)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Точка входа (Root)
// (GET /)
func (_ Unimplemented) GetRoot(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить список пулов
// (GET /pools)
func (_ Unimplemented) ListPools(w http.ResponseWriter, r *http.Request, params ListPoolsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Создание нового пула
// (POST /pools)
func (_ Unimplemented) CreatePool(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Добавить ресурсы в пул
// (POST /pools/{poolId}/resources)
func (_ Unimplemented) AddResources(w http.ResponseWriter, r *http.Request, poolId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetRoot operation middleware
func (siw *ServerInterfaceWrapper) GetRoot(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRoot(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListPools operation middleware
func (siw *ServerInterfaceWrapper) ListPools(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListPoolsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListPools(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreatePool operation middleware
func (siw *ServerInterfaceWrapper) CreatePool(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreatePool(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddResources operation middleware
func (siw *ServerInterfaceWrapper) AddResources(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "poolId" -------------
	var poolId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "poolId", chi.URLParam(r, "poolId"), &poolId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "poolId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddResources(w, r, poolId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/", wrapper.GetRoot)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/pools", wrapper.ListPools)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pools", wrapper.CreatePool)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pools/{poolId}/resources", wrapper.AddResources)
	})

	return r
}

type BadRequestJSONResponse StatusResponse

type NotFoundJSONResponse StatusResponse

type GetRootRequestObject struct {
}

type GetRootResponseObject interface {
	VisitGetRootResponse(w http.ResponseWriter) error
}

type GetRoot200ApplicationHalPlusJSONResponse RootResource

func (response GetRoot200ApplicationHalPlusJSONResponse) VisitGetRootResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/hal+json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListPoolsRequestObject struct {
	Params ListPoolsParams
}

type ListPoolsResponseObject interface {
	VisitListPoolsResponse(w http.ResponseWriter) error
}

type ListPools200ApplicationHalPlusJSONResponse PoolCollectionResponse

func (response ListPools200ApplicationHalPlusJSONResponse) VisitListPoolsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/hal+json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type ListPools400JSONResponse struct{ BadRequestJSONResponse }

func (response ListPools400JSONResponse) VisitListPoolsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreatePoolRequestObject struct {
	Body *CreatePoolJSONRequestBody
}

type CreatePoolResponseObject interface {
	VisitCreatePoolResponse(w http.ResponseWriter) error
}

type CreatePool201ApplicationHalPlusJSONResponse Pool

func (response CreatePool201ApplicationHalPlusJSONResponse) VisitCreatePoolResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/hal+json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreatePool400JSONResponse struct{ BadRequestJSONResponse }

func (response CreatePool400JSONResponse) VisitCreatePoolResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type AddResourcesRequestObject struct {
	PoolId openapi_types.UUID `json:"poolId"`
	Body   *AddResourcesJSONRequestBody
}

type AddResourcesResponseObject interface {
	VisitAddResourcesResponse(w http.ResponseWriter) error
}

type AddResources200ApplicationHalPlusJSONResponse Pool

func (response AddResources200ApplicationHalPlusJSONResponse) VisitAddResourcesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/hal+json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type AddResources400JSONResponse struct{ BadRequestJSONResponse }

func (response AddResources400JSONResponse) VisitAddResourcesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type AddResources404JSONResponse struct{ NotFoundJSONResponse }

func (response AddResources404JSONResponse) VisitAddResourcesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Точка входа (Root)
	// (GET /)
	GetRoot(ctx context.Context, request GetRootRequestObject) (GetRootResponseObject, error)
	// Получить список пулов
	// (GET /pools)
	ListPools(ctx context.Context, request ListPoolsRequestObject) (ListPoolsResponseObject, error)
	// Создание нового пула
	// (POST /pools)
	CreatePool(ctx context.Context, request CreatePoolRequestObject) (CreatePoolResponseObject, error)
	// Добавить ресурсы в пул
	// (POST /pools/{poolId}/resources)
	AddResources(ctx context.Context, request AddResourcesRequestObject) (AddResourcesResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetRoot operation middleware
func (sh *strictHandler) GetRoot(w http.ResponseWriter, r *http.Request) {
	var request GetRootRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetRoot(ctx, request.(GetRootRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetRoot")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetRootResponseObject); ok {
		if err := validResponse.VisitGetRootResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListPools operation middleware
func (sh *strictHandler) ListPools(w http.ResponseWriter, r *http.Request, params ListPoolsParams) {
	var request ListPoolsRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListPools(ctx, request.(ListPoolsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListPools")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListPoolsResponseObject); ok {
		if err := validResponse.VisitListPoolsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreatePool operation middleware
func (sh *strictHandler) CreatePool(w http.ResponseWriter, r *http.Request) {
	var request CreatePoolRequestObject

	var body CreatePoolJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreatePool(ctx, request.(CreatePoolRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreatePool")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreatePoolResponseObject); ok {
		if err := validResponse.VisitCreatePoolResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// AddResources operation middleware
func (sh *strictHandler) AddResources(w http.ResponseWriter, r *http.Request, poolId openapi_types.UUID) {
	var request AddResourcesRequestObject

	request.PoolId = poolId

	var body AddResourcesJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.AddResources(ctx, request.(AddResourcesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "AddResources")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(AddResourcesResponseObject); ok {
		if err := validResponse.VisitAddResourcesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}